[refactor] move config util function to schema package
