+++ pkg/skaffold/initializer/config_test.go
@@ -231,46 +231,6 @@ func TestArtifacts(t *testing.T) {
 	})
 }
 
-func TestIsSkaffoldConfig(t *testing.T) {
-	tests := []struct {
-		description string
-		contents    string
-		isValid     bool
-	}{
-		{
-			description: "valid skaffold config",
-			contents: `apiVersion: skaffold/v1beta6
-kind: Config
-deploy:
-  kustomize: {}`,
-			isValid: true,
-		},
-		{
-			description: "not a valid format",
-			contents:    "test",
-			isValid:     false,
-		},
-		{
-			description: "invalid skaffold config version",
-			contents: `apiVersion: skaffold/v2beta1
-kind: Config
-deploy:
-  kustomize: {}`,
-			isValid: false,
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			tmpDir := t.NewTempDir().
-				Write("skaffold.yaml", test.contents)
-
-			isValid := isSkaffoldConfig(tmpDir.Path("skaffold.yaml"))
-
-			t.CheckDeepEqual(test.isValid, isValid)
-		})
-	}
-}
-
 func Test_canonicalizeName(t *testing.T) {
 	const length253 = "aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa-aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa-aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaaaaaaaa.aaa"
 	tests := []struct {
